+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -9,7 +9,8 @@ CHANGELOG
  * Allowed the `Router` to work with any PSR-11 container
  * Added option in workflow dump command to label graph with a custom label
  * Using a `RouterInterface` that does not implement the `WarmableInterface` is deprecated and will not be supported in Symfony 5.0.
- 
+ * The `RequestDataCollector` class has been deprecated and will be removed in Symfony 5.0. Use the `Symfony\Component\HttpKernel\DataCollector\RequestDataCollector` class instead.
+
 4.0.0
 -----